validator/requirements.txt
upowpy==1.0.7
fastapi==0.110.3
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.0
python-dotenv==1.0.1
requests==2.26.0
slowapi==0.1.9
uvicorn==0.29.0
websockets==12.0
faker==25.2.0
uuid-utils==0.7.0
motor==3.4.0